fru: update fru bus and address type to size_t
sync the type of bus and address with dbus-interfaces's[0] definition,
it's should be size_t, update it for 64-bit systems avoid caller
misuse it.

Tested:
using busctl to get the bus and address the type of bus and address is
`t` for `uint64_t` in aarch64.

```
~# busctl call xyz.openbmc_project.FruDevice /xyz/openbmc_project/FruDevice/FRONT_BP3_FRU org.freedesktop.DBus.Properties Get ss xyz.openbmc_project.FruDevice BUS
v t 19
~# busctl call xyz.openbmc_project.FruDevice /xyz/openbmc_project/FruDevice/FRONT_BP3_FRU org.freedesktop.DBus.Properties Get ss xyz.openbmc_project.FruDevice ADDRESS
v t 84
```

[0]:https://github.com/openbmc/phosphor-dbus-interfaces/blob/master/yaml/xyz/openbmc_project/Inventory/Decorator/I2CDevice.interface.yaml#L7C13-L7C17

+++ src/fru_device/fru_utils.cpp
@@ -1130,7 +1130,7 @@ std::optional<int> findIndexForFRU(
 std::optional<std::string> getProductName(
     std::vector<uint8_t>& device,
     boost::container::flat_map<std::string, std::string>& formattedFRU,
-    uint32_t bus, uint32_t address, size_t& unknownBusObjectCount)
+    size_t bus, size_t address, size_t& unknownBusObjectCount)
 {
     std::string productName;
 
@@ -1170,7 +1170,7 @@ std::optional<std::string> getProductName(
     return productName;
 }
 
-bool getFruData(std::vector<uint8_t>& fruData, uint32_t bus, uint32_t address)
+bool getFruData(std::vector<uint8_t>& fruData, size_t bus, size_t address)
 {
     try
     {
